requirements.txt
flask==3.0.0
werkzeug==3.0.1
pillow==10.1.0
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
gunicorn==21.2.0
requests